Base class
- Moved the logic to generate allowed values for item collections into the base class.
