+++ app/Entity/Item/AllocatedExpense.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 namespace App\Entity\Item;
 
 use App\Request\Hash;
-use App\Request\Parameter\Request;
 use Illuminate\Database\Eloquent\Model;
 use App\Models\Transformers\Transformer;
 use App\Request\Validate\Validator;
@@ -21,34 +20,9 @@ class AllocatedExpense extends Item
         parent::__construct();
     }
 
-    public function allowedValuesForItemCollection(
-        int $resource_type_id,
-        int $resource_id,
-        array $permitted_resource_types = [],
-        bool $include_public = false
-    ): array
+    public function allowedValuesItemCollectionClass(): string
     {
-        $available_parameters = $this->requestParameters();
-        $defined_parameters = Request::fetch(
-            array_keys($available_parameters),
-            $resource_type_id,
-            $resource_id
-        );
-
-        $allowed_values = new \App\Option\AllowedValue\Item\AllocatedExpense(
-            $resource_type_id,
-            $resource_id,
-            $permitted_resource_types,
-            $include_public,
-        );
-
-        return $allowed_values
-            ->setParameters(
-                $available_parameters,
-                $defined_parameters
-            )
-            ->fetch()
-            ->allowedValues();
+        return \App\Option\AllowedValue\Item\AllocatedExpense::class;
     }
 
     /**
