+++ app/Entity/Item/SimpleExpense.php
@@ -5,7 +5,6 @@ namespace App\Entity\Item;
 
 use App\Models\Transformers\Transformer;
 use App\Request\Hash;
-use App\Request\Parameter\Request;
 use App\Request\Validate\Validator;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Date;
@@ -21,34 +20,9 @@ class SimpleExpense extends Item
         parent::__construct();
     }
 
-    public function allowedValuesForItemCollection(
-        int $resource_type_id,
-        int $resource_id,
-        array $permitted_resource_types = [],
-        bool $include_public = false
-    ): array
+    public function allowedValuesItemCollectionClass(): string
     {
-        $available_parameters = array_keys($this->requestParameters());
-        $defined_parameters = Request::fetch(
-            $available_parameters,
-            $resource_type_id,
-            $resource_id
-        );
-
-        $allowed_values = new \App\Option\AllowedValue\Item\SimpleExpense(
-            $resource_type_id,
-            $resource_id,
-            $permitted_resource_types,
-            $include_public,
-        );
-
-        return $allowed_values
-            ->setParameters(
-                $available_parameters,
-                $defined_parameters
-            )
-            ->fetch()
-            ->allowedValues();
+        return \App\Option\AllowedValue\Item\SimpleExpense::class;
     }
 
     public function create(int $id): Model
